checkpoint: moves "*/EndpointResponseError.ts" to "*/Fetching/ResponseError.ts"
